added trap and logio restart params

+++ start_single_worker.sh
@@ -45,9 +45,20 @@ export logio_agent_check_start_time=0
 export LOGIO_AGENT_RESTART_TIME_SECS_THRESHOLD=${LOGIO_AGENT_RESTART_TIME_SECS_THRESHOLD:-300}
 
 export tnow='date +"%Y-%m-%d %H:%M:%S"'
+export worker_pid=0
+export kill_signal=0
 
 function stop_logio_agent() {
-	kill -9 $(ps -eaf | grep log.io- | head -n -1 | awk '{print $2}')
+        #kill -0 $(ps -eaf | grep log.io- | head -n -1 | awk '{print $2}')
+        pgrep -f "node /usr/local/bin/log.io-file-input" > /dev/null
+        status=$?
+        if [ $status == 0 ];then
+                kill -9 $(ps -eaf | grep log.io- | head -n -1 | awk '{print $2}')
+        else
+                echo "Nothing to kill."
+		status=0
+        fi
+
 }
 
 function setup_logio_agent() {
@@ -73,6 +84,8 @@ function setup_logio_agent() {
 	if [ $status == 0 ];then
 		echo "Start input server"
 		nohup log.io-file-input &
+		echo $! > $BASE_PATH/$APP_NAME/logio.pid
+                logio_agent_check_start_time=$(eval $tnow)
 	else
 		echo "Starting logger failed."
 	fi
@@ -171,31 +184,89 @@ function start_workers() {
 }
 
 function restart_logio_agent() {
-    echo "Checking logio agent to restart."
+    echo "Checking logio agent to restart. Last started time $logio_agent_check_start_time."
 
-    if [ $logio_agent_check_start_time == 0 ];then
+    if [ "$logio_agent_check_start_time" == 0 ];then
+	echo "updating check start time."
         logio_agent_check_start_time=$(eval $tnow)
     fi
 
     logio_agent_check_start_time_seconds=$(date -d "$logio_agent_check_start_time" +%s)
-    tnow_secs=$(date -d "$(eval $tnow)" %s)
+    tnow_secs=$(date -d "$(eval $tnow)" +%s)
     tdiff=$((tnow_secs - logio_agent_check_start_time_seconds))
 
-    if [ tdiff > $LOGIO_AGENT_RESTART_TIME_SECS_THRESHOLD ];then
-        echo "Restarting logio agent."
+    if (( $tdiff > $LOGIO_AGENT_RESTART_TIME_SECS_THRESHOLD ));then
+        echo "Restarting logio agent. tdiff is $tdiff, which is higher than set threshold $LOGIO_AGENT_RESTART_TIME_SECS_THRESHOLD."
         setup_logio_agent
         return
     fi
 
-    echo "Not restarting logio agent as threshold - $LOGIO_AGENT_RESTART_TIME_SECS_THRESHOLD seconds not elapsed yet."
+    echo "Not restarting logio agent as threshold - $LOGIO_AGENT_RESTART_TIME_SECS_THRESHOLD seconds not elapsed yet.\n"
 }
 
 function docheckup() {
-    restart_logio_agent
+    if [ $kill_signal == 0 ];then
+	    echo "Running Checks."
+	    restart_logio_agent
+    else
+	    echo "Kill received, skipping checks."
+	    return 
+    fi
+}
+
+function is_process_running() {
+    ssw_pid=$1
+    if ! kill -0 $ssw_pid > /dev/null 2>&1; then
+        echo "Process $ssw_pid is not running." 1>&2
+        return 1
+    fi
+    return 0
+}
+
+function kill_process() {
+    # Kill the process with the given PID and a timeout of TERMINATION_GRACE_PERIOD or 130 seconds
+    ssw_pid=$(cat ./worker.pid)
+    shutdown_time=${2:-${TERMINATION_GRACE_PERIOD}}
+    echo "Killing process $ssw_pid with a timeout of $shutdown_time seconds" 1>&2
+    if ! is_process_running $ssw_pid; then
+        return 0
+    fi
+    kill -15 $ssw_pid > /dev/null 2>&1 # Send SIGTERM to the process
+    echo "Waiting for process $ssw_pid to terminate, with a timeout of $shutdown_time seconds..." 1>&2
+    for i in $(seq 1 $shutdown_time); do
+        if ! is_process_running $ssw_pid; then
+            echo "Process $ssw_pid terminated successfully." 1>&2
+            return 0
+        fi
+        sleep 1
+        echo "Process $ssw_pid is still running. Waiting since $i seconds..." 1>&2
+    done
+    kill -9 $ssw_pid > /dev/null 2>&1 # Send SIGKILL to the process
+    sleep 2
+    if ! is_process_running $ssw_pid; then
+        echo "Process $ssw_pid forcefully terminated." 1>&2
+        return 0
+    fi
+    echo "Failed to terminate process $ssw_pid."
+    return 1
+}
+
+function stop_single_workers() {
+    export kill_signal=1
+    stop_logio_agent
+    shutdown_time=${1:-${TERMINATION_GRACE_PERIOD}}
+    pids=()
+    is_success=0
+    echo "Stopping workers $ENTRYPOINT_PREFIX worker with a timeout of $shutdown_time seconds" 1>&2
+    kill_process 
+
+    
 }
 
 function main() {
-	echo "Starting ${SERVICE_NAME} in BG. Logs are written to stdout > $STDOUT_LOG_FILE err > $STDERR_LOG_FILE app > $LOG_FILE"
+    trap "echo 'Received kill signal'; stop_single_workers" SIGTERM SIGINT SIGHUP
+    echo "Trap is set."
+    echo "Starting ${SERVICE_NAME} in BG. Logs are written to stdout > $STDOUT_LOG_FILE err > $STDERR_LOG_FILE app > $LOG_FILE"
     if [ $PYTHON_VENV != 0 ];then
     	export WAITRESS_PATH=$PYTHON_VENV/bin/waitress-serve
     	export WORKER_PATH=$PYTHON_VENV/bin/worker
@@ -214,6 +285,7 @@ function main() {
 
 	    app_pid=$!
 		echo $app_pid > app.pid
+        echo $app_pid > worker.pid
         if [ $SETUP_LOGIO_AGENT == "True" ];then
             sleep 5
             setup_logio_agent
@@ -239,6 +311,7 @@ function main() {
 			w_pid=$!
 			echo "PID is $w_pid"
 			echo $w_pid > $a.pid
+            echo $w_pid > worker.pid
             if [ $SETUP_LOGIO_AGENT == "True" ];then
                 sleep 5
                 setup_logio_agent
@@ -276,6 +349,7 @@ function main() {
 			w_pid=$!
 			echo "PID is $w_pid"
 			echo $w_pid > $a.pid
+            echo $w_pid > worker.pid
             if [ $SETUP_LOGIO_AGENT == "True" ];then
                 sleep 5
                 setup_logio_agent
